# executor: add daily test for information_schema.tables