--- a/pkg/executor/union_scan_test.go
+++ b/pkg/executor/union_scan_test.go
@@ -479,5 +479,6 @@ func TestBenchDaily(t *testing.T) {
 		BenchmarkUnionScanIndexReadDescRead,
 		BenchmarkUnionScanTableReadDescRead,
 		BenchmarkUnionScanIndexLookUpDescRead,
+		BenchmarkInfoschemaTables,
 	)
 }