<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: ref #62020

### What changed and how does it work?

The performance regression happened silently. Let's add a daily benchmark to detect the performance regression in time.

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
  Before #62042:
  ```
  goos: linux
  goarch: amd64
  pkg: github.com/pingcap/tidb/pkg/executor
  cpu: 13th Gen Intel(R) Core(TM) i9-13900K
  BenchmarkInfoschemaTables
  BenchmarkInfoschemaTables/select_all
  BenchmarkInfoschemaTables/select_all-32         	     451	   2778437 ns/op	 1382168 B/op	   15495 allocs/op
  BenchmarkInfoschemaTables/select_basic
  BenchmarkInfoschemaTables/select_basic-32       	    6283	    187001 ns/op	   99893 B/op	    1153 allocs/op
  BenchmarkInfoschemaTables/select_without_table_rows
  BenchmarkInfoschemaTables/select_without_table_rows-32         	    3922	    315444 ns/op	  185616 B/op	    2032 allocs/op
  BenchmarkInfoschemaTables/select_with_table_rows
  BenchmarkInfoschemaTables/select_with_table_rows-32            	     442	   2732047 ns/op	 1165547 B/op	   13513 allocs/op
  PASS
  ok  	github.com/pingcap/tidb/pkg/executor	7.060s
  ```
  After #62042:
  ```
  goos: linux
  goarch: amd64
  pkg: github.com/pingcap/tidb/pkg/executor
  cpu: 13th Gen Intel(R) Core(TM) i9-13900K
  BenchmarkInfoschemaTables
  BenchmarkInfoschemaTables/select_all
  BenchmarkInfoschemaTables/select_all-32         	     951	   1346347 ns/op	  752420 B/op	    8215 allocs/op
  BenchmarkInfoschemaTables/select_basic
  BenchmarkInfoschemaTables/select_basic-32       	    6516	    196947 ns/op	  100023 B/op	    1153 allocs/op
  BenchmarkInfoschemaTables/select_without_table_rows
  BenchmarkInfoschemaTables/select_without_table_rows-32         	    3066	    345877 ns/op	  185584 B/op	    2036 allocs/op
  BenchmarkInfoschemaTables/select_with_table_rows
  BenchmarkInfoschemaTables/select_with_table_rows-32            	     974	   1264975 ns/op	  535432 B/op	    6231 allocs/op
  PASS
  ok  	github.com/pingcap/tidb/pkg/executor	5.790s
  ```
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```